Add: Additional layer of security for admin functionality

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -71,8 +72,14 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     "Allows super user to add products to the store"
+
+    if not request.user.is_superuser:
+        messages.error(request, 'This functionality is reserved for admins!')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
 
@@ -82,7 +89,7 @@ def add_product(request):
             return redirect(reverse('product_detail', args=[product.id]))
         else:
             messages.error(request, "Whoops! There seems to be an error in your form.")
-            
+
     else:
         form = ProductForm()
 
@@ -93,8 +100,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request):
     "Allows super user to edit products in the store"
+
+    if not request.user.is_superuser:
+        messages.error(request, 'This functionality is reserved for admins!')
+        return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
 
@@ -118,8 +130,13 @@ def edit_product(request):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """Allows super user to delete products from store"""
+
+    if not request.user.is_superuser:
+        messages.error(request, 'This functionality is reserved for admins!')
+        return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
